rename app-google.js to googleFilePicker.js
