+++ src/fileLoad.js
@@ -1,5 +1,5 @@
 import * as Utils from './utils.js';
-import * as GoogleFilePicker from './app-google.js';
+import * as GoogleFilePicker from './googleFilePicker.js';
 import { DOMUtils } from '../node_modules/igv-ui/src/index.js'
 
 class FileLoad {
