+++ src/index.js
@@ -5,7 +5,7 @@ import FileLoad from "./fileLoad.js";
 import GenomeFileLoad from "./genomeFileLoad.js";
 import SessionFileLoad from "./sessionFileLoad.js";
 import TrackFileLoad from "./trackFileLoad.js";
-import * as GoogleFilePicker from './app-google.js';
+import * as GoogleFilePicker from './googleFilePicker.js';
 import * as Utils from './utils.js';
